refactor(auth): merge user module into auth
User utilities, schema, hooks, and RPC belong with authentication — consolidate apps/user into apps/auth and update imports.
