+++ apps/auth/index.ts
@@ -1,2 +1,4 @@
 export * from "./auth.server"
+export * from "./hooks"
 export * from "./rpc"
+export * from "./user.server"
