+++ apps/auth/rpc/guards.ts
@@ -1,8 +1,8 @@
 import { redirect } from "@tanstack/react-router"
 import { createServerFn } from "@tanstack/react-start"
 import { getRequestUrl } from "@tanstack/react-start/server"
-import { getUserId, isValidUserId } from "@/user"
 import { useAppSession } from "../session.server"
+import { getUserId, isValidUserId } from "../user.server"
 
 const LOGIN_ROUTE = "/login"
