+++ apps/auth/rpc/index.ts
@@ -1,3 +1,4 @@
+export * from "./get-user"
 export * from "./guards"
 export * from "./login"
 export * from "./logout"
