+++ apps/permissions/rpc/guards.test.ts
@@ -11,9 +11,9 @@ import {
   it,
   vi,
 } from "vitest"
+import { users } from "@/auth/db/schema"
 import { getDb } from "@/db"
 import * as schema from "@/db/schema"
-import { users } from "@/user/db/schema"
 import { userRoles } from "../db/schema"
 import type { Permission, PermissionsConfig, Role } from "../permissions.server"
 import { requireUserPermissions, requireUserRole } from "./guards"
