--- apps/user/index.ts
@@ -1,3 +1,0 @@
-export * from "./hooks"
-export * from "./rpc"
-export * from "./user.server"
