--- apps/user/rpc/index.ts
@@ -1,1 +1,0 @@
-export * from "./get-user"
