+++ packages/db/schema.ts
@@ -1,5 +1,5 @@
 // File generated by bun run db:generate:imports
 
+export * from "@/auth/db/schema"
 export * from "@/content/db/schema"
 export * from "@/permissions/db/schema"
-export * from "@/user/db/schema"
